DjRest/requirements.txt
Django>=2.2.12
psycopg2-binary>=2.8
djangorestframework>=3.12.4